requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.13
wikipedia==1.4.0
